// back-end/src/types/roles.ts
export enum Role {
  ADMINISTRADOR = 'ADMINISTRADOR',
  RECEPCIONISTA = 'RECEPCIONISTA',
  CORTADOR = 'CORTADOR',
  DISEÑADOR = 'DISEÑADOR',
  REPRESENTANTE_TALLER = 'REPRESENTANTE_TALLER',
  AYUDANTE = 'AYUDANTE'
}

export enum Permission {
  // Usuarios
  VIEW_USERS = 'VIEW_USERS',
  MANAGE_USERS = 'MANAGE_USERS',
  DELETE_USERS = 'DELETE_USERS',
  UPDATE_OWN_PROFILE = 'UPDATE_OWN_PROFILE',
  
  // Clientes
  VIEW_CLIENTS = 'VIEW_CLIENTS',
  MANAGE_CLIENTS = 'MANAGE_CLIENTS',
  
  // Productos
  VIEW_PRODUCTS = 'VIEW_PRODUCTS',
  MANAGE_PRODUCTS = 'MANAGE_PRODUCTS',
  
  // Inventario
  VIEW_INVENTORY = 'VIEW_INVENTORY',
  MANAGE_INVENTORY = 'MANAGE_INVENTORY',
  REGISTER_MATERIAL_USAGE = 'REGISTER_MATERIAL_USAGE',
  
  // Pedidos
  VIEW_ORDERS = 'VIEW_ORDERS',
  MANAGE_ORDERS = 'MANAGE_ORDERS',
  
  // Cotizaciones
  VIEW_QUOTES = 'VIEW_QUOTES',
  MANAGE_QUOTES = 'MANAGE_QUOTES',
  
  // Confecciones
  VIEW_CONFECTIONS = 'VIEW_CONFECTIONS',
  MANAGE_CONFECTIONS = 'MANAGE_CONFECTIONS',
  
  // Talleres
  VIEW_WORKSHOPS = 'VIEW_WORKSHOPS',
  MANAGE_WORKSHOPS = 'MANAGE_WORKSHOPS',
  UPDATE_WORKSHOP_STATUS = 'UPDATE_WORKSHOP_STATUS',
  
  // Despachos
  VIEW_DISPATCH = 'VIEW_DISPATCH',
  MANAGE_DISPATCH = 'MANAGE_DISPATCH',
  CONFIRM_DELIVERY = 'CONFIRM_DELIVERY',
  
  // Dashboard
  VIEW_DASHBOARD = 'VIEW_DASHBOARD',
  VIEW_REPORTS = 'VIEW_REPORTS',
}